Add polyfills to demo
This allows browser crosscompatibility. Not all browsers fully support web components and polymer, polyfills fills up the gaps. It is an ever-shrinking library that may at some point be removed.

Stuck it in the demo rather than the component, because the component is fully compliant and shouldn't import legacy support fills.